+++ src/app/components/Header/UserAccountInfo.tsx
@@ -16,7 +16,7 @@ const UserAccountInfo = ({ email, profileImgSrc }: UserAccountInfoProps) => {
           src={profileImgSrc ?? DEFAULT_PROFILE_IMG_SRC}
           alt="profile image"
           fill
-          sizes="10vw"
+          sizes="64vw"
         />
       </div>
       <div className={styles.usernameContainer}>{email}</div>
