+++ src/app/components/LinkCardList/LinkCard/LinkCard.tsx
@@ -143,7 +143,7 @@ const LinkCard = ({
             src={thumbnailSrc ?? "/default-thumbnail.svg"}
             alt="thumbnail"
             fill
-            sizes="10vw"
+            sizes="64vw"
           />
           <div
             className={styles.likeBtnBox}
@@ -156,13 +156,13 @@ const LinkCard = ({
               src={isLiked ? "/like-btn-liked.svg" : "/like-btn-unliked.svg"}
               alt="like button"
               fill
-              sizes="1vw"
+              sizes="32vw"
             />
           </div>
         </div>
         <div className={styles.metadataContainer}>
           <div className={styles.kebab} onClick={handleClickOpenPopover}>
-            <Image src="/kebab.svg" alt="kebab" fill sizes="1vw" />
+            <Image src="/kebab.svg" alt="kebab" fill sizes="32vw" />
             {showPopover && (
               <div className={styles.popoverContainer} ref={popoverRef}>
                 <Popover
