+++ src/app/components/Loader/Loader.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import styles from "./Loader.module.scss";
+
+const Loader = () => {
+  return (
+    <div className={styles.loader}>
+      <div className={styles.loading}></div>
+    </div>
+  );
+};
+
+export default Loader;
