+++ src/app/folder/[[...id]]/AddLink/AddLink.module.scss
