+++ src/app/folder/[[...id]]/FolderChip/FolderChip.module.scss
