+++ src/app/folder/[[...id]]/Option/Option.module.scss
