+++ src/app/folder/[[...id]]/layout.tsx
@@ -1,0 +1,14 @@
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { getServerSession } from "next-auth";
+import { ReactNode } from "react";
+import { redirect } from "next/navigation";
+
+const layout = async ({ children }: { children: ReactNode }) => {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    redirect("/signin");
+  }
+  return <>{children}</>;
+};
+
+export default layout;
