+++ src/app/folder/[[...id]]/page.module.scss
@@ -1,0 +1,64 @@
+@use "/src/styles/variables" as var;
+
+.mainContainer {
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  transition: all 0.3s ease;
+  &.floating {
+    flex-direction: column-reverse;
+  }
+}
+
+.heroSection {
+  padding: 60px 32.5px 90px;
+  background: var.$gray-1;
+  transition: all 0.3s ease;
+  &.floating {
+    position: -webkit-sticky;
+    position: sticky;
+    bottom: 0;
+    padding: 24px 32.5px;
+    width: 100%;
+    z-index: 11;
+    @media only screen and (max-width: 768px) {
+      padding: 16px 32.5px;
+    }
+  }
+  @media only screen and (max-width: 768px) {
+    padding: 24px 32.5px 40px;
+  }
+
+  .addLinkContainer {
+    width: min(100%, 800px);
+    margin: 0 auto;
+  }
+}
+
+.contentSection {
+  padding: 40px 32.5px 100px;
+
+  @media only screen and (max-width: 768px) {
+    padding: 20px 32.5px 60px;
+  }
+
+  .contentsContainer {
+    max-width: 1060px;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    padding: 0px;
+    gap: 40px;
+    margin: 0 auto;
+
+    .searchBarContainer {
+      width: 100%;
+    }
+  }
+}
+
+.thresholdTarget {
+  width: 100%;
+  height: 1px;
+  background-color: var.$white;
+}
