--- src/app/folder/page.module.scss
@@ -1,170 +1,0 @@
-@use "/src/styles/variables" as var;
-
-.mainContainer {
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  transition: all 0.3s ease;
-  &.floating {
-    flex-direction: column-reverse;
-  }
-}
-
-.heroSection {
-  padding: 60px 32.5px 90px;
-  background: var.$gray-1;
-  transition: all 0.3s ease;
-  &.floating {
-    position: -webkit-sticky;
-    position: sticky;
-    bottom: 0;
-    padding: 24px 32.5px;
-    width: 100%;
-    z-index: 11;
-    @media only screen and (max-width: 768px) {
-      padding: 16px 32.5px;
-    }
-  }
-  @media only screen and (max-width: 768px) {
-    padding: 24px 32.5px 40px;
-  }
-
-  .addLinkContainer {
-    width: min(100%, 800px);
-    margin: 0 auto;
-  }
-}
-
-.contentSection {
-  padding: 40px 32.5px 100px;
-
-  @media only screen and (max-width: 768px) {
-    padding: 20px 32.5px 60px;
-  }
-
-  .contentsContainer {
-    max-width: 1060px;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    padding: 0px;
-    gap: 40px;
-    margin: 0 auto;
-
-    .searchBarContainer {
-      width: 100%;
-    }
-
-    .mainContentContainer {
-      display: flex;
-      width: 100%;
-      flex-direction: column;
-      padding: 0px;
-      gap: 24px;
-
-      .chipListRow {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        align-items: center;
-        padding: 0px;
-
-        .chipListContainer {
-          display: flex;
-          padding: 0px;
-          gap: 8px;
-          list-style-type: none;
-          flex-wrap: wrap;
-        }
-
-        .addFolderButton {
-          background: none;
-          color: var.$primary;
-          font-weight: 500;
-          display: flex;
-          align-items: center;
-          gap: 4px;
-          padding: 0px;
-          cursor: pointer;
-
-          transition: all 0.2s ease;
-          &:hover {
-            font-weight: 600;
-            transform: scale(1.2);
-          }
-          @media only screen and (max-width: 768px) {
-            position: fixed;
-            bottom: 100px;
-            left: calc(50% - 128px / 2);
-            background-color: var.$primary;
-            color: var.$white;
-            width: 128px;
-            height: 36px;
-            justify-content: center;
-            border-radius: 20px;
-            z-index: 10;
-            &:hover {
-              transform: scale(1.05);
-            }
-          }
-          span {
-            line-height: 0%;
-          }
-          .addIconBox {
-            position: relative;
-            display: inline-block;
-            width: 16px;
-            height: 16px;
-            & > img:nth-child(2) {
-              display: none;
-            }
-            @media only screen and (max-width: 786px) {
-              & > img:first-child {
-                display: none;
-              }
-              & > img:nth-child(2) {
-                display: block;
-              }
-            }
-          }
-        }
-      }
-
-      .titleRow {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        align-items: center;
-        padding: 0px;
-
-        @media only screen and (max-width: 768px) {
-          flex-direction: column;
-          align-items: flex-start;
-          gap: 12px;
-        }
-        .title {
-          font-weight: 600;
-          line-height: 28.64px;
-        }
-
-        .optionListContainer {
-          display: flex;
-          align-items: flex-start;
-          list-style: none;
-          padding: 0px;
-          gap: 12px;
-        }
-      }
-
-      .cardList {
-        margin: 0 auto;
-      }
-    }
-  }
-}
-
-.thresholdTarget {
-  width: 100%;
-  height: 1px;
-  background-color: var.$white;
-}
