+++ src/app/types/next-auth.d.ts
@@ -1,0 +1,15 @@
+import NextAuth, { DefaultSession } from "next-auth";
+
+declare module "next-auth" {
+  /**
+   * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
+   */
+  interface Session {
+    user: {
+      id?: number;
+      name?: string;
+      email?: string;
+      image?: string;
+    } & DefaultSession["user"];
+  }
+}
