+++ src/app/types/types.d.ts
@@ -1,0 +1,6 @@
+export interface Folder {
+  id: number;
+  created_at: string;
+  name: string;
+  user_id: number;
+}
