+++ src/lib/api/instance.ts
@@ -13,7 +13,7 @@ export const wrapRequest = (func: any) => {
 };
 
 export const instance = axios.create({
-  baseURL: "https://bootcamp-api.codeit.kr/api",
+  baseURL: "http://localhost:3000/linkbrary",
   timeout: 10000,
   headers: {
     "Content-Type": "application/json",
